chore: move ledger app hook into hooks directory
