+++ src/components/MultiHopTrade/components/VerifyAddresses/VerifyAddresses.tsx
@@ -26,6 +26,7 @@ import { SlideTransition } from 'components/SlideTransition'
 import { RawText, Text } from 'components/Text'
 import type { TextPropTypes } from 'components/Text/Text'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
+import { useLedgerOpenApp } from 'hooks/useLedgerOpenApp/useLedgerOpenApp'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { walletSupportsChain } from 'hooks/useWalletSupportsChain/useWalletSupportsChain'
 import {
@@ -37,7 +38,6 @@ import {
 } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
-import { useLedgerOpenApp } from '../MultiHopTradeConfirm/components/useLedgerOpenApp'
 import { WithBackButton } from '../WithBackButton'
 
 export const VerifyAddresses = () => {
